move QgsTriangle tests from QgsGeometry
